requirements.txt
torch
streamlit
Pillow
transformers
openai
python-dotenv
sentence-transformers
faiss-cpu
torch
google-search-results
easyocr
flask
werkzeug
pillow